rayserve/requirements.txt
ray[serve]==2.6.0
fastapi==0.95.2
pydantic==1.10.11
numpy
Pillow
